==> 08-Figure_7.Rmd <==
---
title: "Figure 6 - In Vitro Differentiation"
output: html_document
editor_options: 
  chunk_output_type: console
creative_commons: CC BY
---


# Figure 6A -  Dmrt1 and Sox9 expression timecourse
```{r Dmrt1_Sox9_invitro_timecourse,eval=T}
as.data.frame(colData(dds_invitro<-dds[,!str_detect(colData(dds)$time,"ovary|testis") & !str_detect(colData(dds)$batch,"GW") ]))
dds_invitro<-estimateSizeFactors(dds_invitro)

svglite::svglite(paste0("vitro_Dmrt1_expression_",ts,".svg"),width=5,height=3)
gg_plotCounts("Dmrt1",d=dds_invitro,fpkm=T,bar=T)
dev.off()

svglite::svglite(paste0("vitro_Sox9_expression_",ts,".svg"),width=5,height=3)
gg_plotCounts("Sox9",d=dds_invitro,fpkm=T,bar=T)
dev.off()

svglite::svglite(paste0("vitro_Plppr4_expression_",ts,".svg"),width=5,height=3)
gg_plotCounts("Plppr4",d=dds_invitro,fpkm=T,bar=T)
dev.off()


```


# Figure 6B - In vitro expression PCA

```{r invitro_PCA,eval=T}

colors<-c(cbPalette[8],cbPalette[6],c1a,c1b,c1c,c1d,c2a,c2b,c2c,c2d,c3a,c3b,c3c,c3d)
#plotPCA(normTransform(dds_invitro),intgroup=c("gene","time"),ntop = 500,returnData=F)

figure_6b<-plotPCA(normTransform(dds_invitro),intgroup=c("gene","time"),ntop = 500,returnData=T) %>% 
  mutate(time=factor(time,levels=c("fresh","0h","24h","48h","7d"))) %>% 
  mutate(group=factor(group,levels=c("control:fresh","sertoli:fresh",
                                     "control:0h","control:24h", "control:48h", "control:7d",
                                     "sox9:0h","sox9:24h","sox9:48h","sox9:7d",
                                     "dmrt1:0h","dmrt1:24h","dmrt1:48h","dmrt1:7d"))) %>% 
  ggplot(aes(x=PC1,y=PC2,fill=group,shape=time)) + geom_point(size=8) +
  xlim(c(-45,60)) + ylim(c(-100,40)) +
  ggtitle("In Vitro Data") +
  scale_fill_manual(values = colors)+  scale_shape_manual(values=c(21,21,23,22,24)) +
  theme_bw() + theme(legend.position = "none")
 
svglite::svglite(paste0("Figure6b_in_vitro_PCA_",ts,".svg"),width=5,height=4.5)
figure_6b
dev.off()
```


```{r}
#Export version without Log2 for Table
dds_invitro_full<-dds[,!str_detect(colData(dds)$time,"ovary|testis") ]

f_invitro_mean_symbol <- as.data.frame(fpkm(dds_invitro_full)) %>% 
#  dplyr::filter(mgi_symbol %in% c(goi,goi2)) %>% 
  tibble::rownames_to_column(var="ensembl") %>%
  tidyr::gather(sample,fpkm,-ensembl) %>%
  dplyr::mutate(time=colData(dds_invitro_full)[sample,]$time) %>%
  dplyr::mutate(gene=colData(dds_invitro_full)[sample,]$gene) %>%
  dplyr::mutate(sample=paste(time,gene,sep="_")) %>% 
  dplyr::group_by(sample,ensembl) %>% 
    dplyr::summarize(mean_fpkm=round(mean(fpkm),3)) %>% 
    dplyr::select(sample,ensembl,mean_fpkm) %>% 
    ungroup() %>% 
  tidyr::spread(sample,mean_fpkm) %>% as.data.frame()

f_invitro_mean_symbol<-f_invitro_mean_symbol[apply(f_invitro_mean_symbol[,2:17],1,sum) > 0,]
f_invitro_mean_symbol$Symbol<-symbols[match(f_invitro_mean_symbol$ensembl,symbols$gene_id),]$gene_name

f_invitro_mean_symbol <- dplyr::rename(f_invitro_mean_symbol,
                                        "Ensembl GeneID" = "ensembl",
                                        "Freshly Isolated Granulosa Cells" = "fresh_control",
                                        "Freshly Isolated Sertoli Cells" = "fresh_sertoli",
                                        "Control 0 DPTx" = "0h_control",
                                        "Control 1 DPTx" = "24h_control",
                                        "Control 2 DPTx" = "48h_control",
                                        "Control 6 DPTx" = "7d_control",
                                        "CAG-Sox9 0 DPTx" = "0h_sox9",
                                        "CAG-Sox9 1 DPTx" = "24h_sox9",
                                        "CAG-Sox9 2 DPTx" = "48h_sox9",
                                        "CAG-Sox9 6 DPTx" = "7d_sox9",
                                        "CAG-Dmrt1 0 DPTx" = "0h_dmrt1",
                                        "CAG-Dmrt1 1 DPTx" = "24h_dmrt1",
                                        "CAG-Dmrt1 2 DPTx" = "48h_dmrt1",
                                        "CAG-Dmrt1 6 DPTx" = "7d_dmrt1",
                                        "Sox8-/-;Sox9+/-;CAG-Dmrt1 6 DPTx" = "7d_S01_dmrt1",
                                        "Sox8-/-;Sox9-/-;CAG-Dmrt1 6 DPTx" = "7d_S00_dmrt1") %>%
  dplyr::select("Ensembl GeneID","Symbol", "Freshly Isolated Granulosa Cells","Freshly Isolated Sertoli Cells",
                "Control 0 DPTx", "Control 1 DPTx", "Control 2 DPTx", "Control 6 DPTx",
                "CAG-Sox9 0 DPTx", "CAG-Sox9 1 DPTx", "CAG-Sox9 2 DPTx", "CAG-Sox9 6 DPTx",
                "CAG-Dmrt1 0 DPTx", "CAG-Dmrt1 1 DPTx", "CAG-Dmrt1 2 DPTx", "CAG-Dmrt1 6 DPTx",
                "Sox8-/-;Sox9+/-;CAG-Dmrt1 6 DPTx", "Sox8-/-;Sox9-/-;CAG-Dmrt1 6 DPTx")

#View(f_invitro_mean_symbol)
write.csv(f_invitro_mean_symbol,"Lindeman_et_al_Mean_FPKMs_for_Granulosa_Sertoli_and_Cultured_Granulosa_Cells.csv",row.names=F,quote=F)
dim(f_invitro)
```


## Day 7 Gene expression Differences
```{r DESEQ_7D_invitro,eval=T}
as.data.frame(colData(dds_7d<-dds_invitro[,str_detect(colData(dds_invitro)$time,"7d")]))
#as.data.frame(colData(dds_invitro)) #fresh samples included
design(dds_7d)<- ~ gene
dds_7d$gene <- droplevels(dds_7d$gene)
dds_7d<-DESeq(dds_7d)
resultsNames(dds_7d)

summary(res_7d_full_dmrt1<-results(dds_7d, contrast=c("gene","dmrt1","control"),alpha=0.05))
res_7d_full_dmrt1$symbol<-symbols[match(rownames(res_7d_full_dmrt1),symbols$gene_id),]$gene_name
res_7d_full_dmrt1$chr<-symbols[match(rownames(res_7d_full_dmrt1),symbols$gene_id),]$chr


summary(res_7d_full_sox9<-results(dds_7d, contrast=c("gene","sox9","control"),alpha=0.05))
res_7d_full_sox9$symbol<-symbols[match(rownames(res_7d_full_sox9),symbols$gene_id),]$gene_name
res_7d_full_sox9$chr<-symbols[match(rownames(res_7d_full_sox9),symbols$gene_id),]$chr


```

# Figure 6C -GSEA
```{r export_GSEA_lists,eval=T}
somatic_gene_list<-list(sertoli=sertoli_up, granulosa=granulosa_up)
names(somatic_gene_list)

nm<-lapply(seq_along(somatic_gene_list), function(i) c(names(somatic_gene_list)[[i]],"NA", somatic_gene_list[[i]]))

sink("somatic_gene_list.gmt")
writeLines(unlist(lapply(nm, paste, collapse="\t")))
sink()

dim(temp<-as.data.frame(subset(res_7d_full_dmrt1,baseMean > 0 & padj < 0.05) ))
temp$ensembl<-rownames(temp)
temp<-temp[!is.na(temp$stat),]
dim(temp)
#write.table(temp[,c("ensembl","stat")],file="res_7d_full_dmrt1.rnk",sep="\t",quote=F,col.names=F,row.names = F)

#x<-read.table("dmrt1.GseaPreranked.1580504038711/ranked_gene_list_na_pos_versus_na_neg_1580504038711.xls",header=T)
#xnot<-rownames(temp)[!rownames(temp) %in% x$NAME]
#length(xnot<-x$NAME[!x$NAME %in% rownames(temp)])
#View(as.data.frame(res_7d_full_dmrt1[xnot,]))

table(temp$class<-case_when(
temp$ensembl %in% sertoli_up ~ "Sertoli-biased Gene",
temp$ensembl %in% granulosa_up ~ "Granulosa-biased Gene",
TRUE ~ " "))


temp3<-left_join(rownames_to_column(temp,var="Ensembl GeneID"),f_invitro_mean_symbol) %>% 
  arrange(stat) %>% 
  dplyr::filter(`Control 6 DPTx` > 2.5 | `CAG-Dmrt1 6 DPTx` > 2.5) %>%
  dplyr::rename( "Gene Symbol" = "symbol",
                 "Mean Counts" = "baseMean",
                 "Log2FoldChange" = "log2FoldChange",
                 "BH padj" = "padj",
                 "t statistic" = "stat",
                 "Sex Bias" = "class",
                 "FPKM Control 0 DPTx" = "Control 0 DPTx",
                 "FPKM Control 1 DPTx" = "Control 1 DPTx",
                 "FPKM Control 2 DPTx" = "Control 2 DPTx",
                 "FPKM Control 6 DPTx" = "Control 6 DPTx",
                "FPKM CAG-Dmrt1 0 DPTx" = "CAG-Dmrt1 0 DPTx",
                "FPKM CAG-Dmrt1 1 DPTx" = "CAG-Dmrt1 1 DPTx",
                "FPKM CAG-Dmrt1 2 DPTx" = "CAG-Dmrt1 2 DPTx",
                "FPKM CAG-Dmrt1 6 DPTx" = "CAG-Dmrt1 6 DPTx") %>%
  dplyr::select("Ensembl GeneID","Gene Symbol","Mean Counts","Log2FoldChange","BH padj","t statistic","Sex Bias",
                 "FPKM Control 0 DPTx", "FPKM Control 1 DPTx", "FPKM Control 2 DPTx", "FPKM Control 6 DPTx",
                "FPKM CAG-Dmrt1 0 DPTx", "FPKM CAG-Dmrt1 1 DPTx", "FPKM CAG-Dmrt1 2 DPTx", "FPKM CAG-Dmrt1 6 DPTx")
dim(temp3)
write.csv(temp3,"Lindeman_et_al_CagDmrt1_cultured_granulosa.csv",row.names=F,quote=F)
write.table(temp3[,c("Ensembl GeneID","t statistic")],file=paste0("res_7d_full_cagdmrt1_",ts,".rnk"),sep="\t",quote=F,col.names=F,row.names = F)



temp<-as.data.frame(subset(res_7d_full_sox9,baseMean > 0 & padj < 0.05) )
temp$ensembl<-rownames(temp)
temp<-temp[!is.na(temp$stat),]
#write.table(temp[,c("ensembl","stat")],file="res_7d_full_sox9.rnk",sep="\t",quote=F,col.names=F,row.names = F)

table(temp$class<-case_when(
temp$ensembl %in% sertoli_up ~ "Sertoli-biased Gene",
temp$ensembl %in% granulosa_up ~ "Granulosa-biased Gene",
TRUE ~ " "))

temp4<-left_join(rownames_to_column(temp,var="Ensembl GeneID"),f_invitro_mean_symbol) %>% 
  arrange(stat) %>% 
  dplyr::filter(`Control 6 DPTx` > 2.5 | `CAG-Sox9 6 DPTx` > 2.5) %>%
  dplyr::rename( "Gene Symbol" = "symbol",
                 "Mean Counts" = "baseMean",
                 "Log2FoldChange" = "log2FoldChange",
                 "BH padj" = "padj",
                 "t statistic" = "stat",
                  "Sex Bias" = "class",
                 "FPKM Control 0 DPTx" = "Control 0 DPTx",
                 "FPKM Control 1 DPTx" = "Control 1 DPTx",
                 "FPKM Control 2 DPTx" = "Control 2 DPTx",
                 "FPKM Control 6 DPTx" = "Control 6 DPTx",
                "FPKM CAG-Sox9 0 DPTx" = "CAG-Sox9 0 DPTx",
                "FPKM CAG-Sox9 1 DPTx" = "CAG-Sox9 1 DPTx",
                "FPKM CAG-Sox9 2 DPTx" = "CAG-Sox9 2 DPTx",
                "FPKM CAG-Sox9 6 DPTx" = "CAG-Sox9 6 DPTx") %>%
  dplyr::select("Ensembl GeneID","Gene Symbol","Mean Counts","Log2FoldChange","BH padj","t statistic", "Sex Bias",
                 "FPKM Control 0 DPTx", "FPKM Control 1 DPTx", "FPKM Control 2 DPTx", "FPKM Control 6 DPTx",
                "FPKM CAG-Sox9 0 DPTx", "FPKM CAG-Sox9 1 DPTx", "FPKM CAG-Sox9 2 DPTx", "FPKM CAG-Sox9 6 DPTx")
dim(temp4)
write.csv(temp4,"Lindeman_et_al_CagSox9_cultured_granulosa.csv",row.names=F,quote=F)
write.table(temp4[,c("Ensembl GeneID","t statistic")],file=paste0("res_7d_full_cagsox9_",ts,".rnk"),sep="\t",quote=F,col.names=F,row.names = F)


```

# Run Command line version of GSEA
```{r command_line_gsea,eval=F}

#gsea-cli.sh GSEAPreranked -gmx /Users/gearh006/Documents/Notebook/GCD Consultant/zarkowe0/granulosa/somatic_gene_list.gmt -collapse No_Collapse -mode Max_probe -norm meandiv -nperm 1000 -rnk /Users/gearh006/Documents/Notebook/GCD Consultant/zarkowe0/granulosa/res_7d_full_dmrt1.rnk -scoring_scheme classic -rpt_label dmrt1 -create_svgs true -include_only_symbols true -make_sets true -plot_top_x 20 -rnd_seed timestamp -set_max 500 -set_min 15 -zip_report false -out /Users/gearh006/Documents/Notebook/GCD Consultant/zarkowe0/granulosa

#gsea-cli.sh GSEAPreranked -gmx /Users/gearh006/Documents/Notebook/GCD Consultant/zarkowe0/granulosa/somatic_gene_list.gmt -collapse No_Collapse -mode Max_probe -norm meandiv -nperm 1000 -rnk /Users/gearh006/Documents/Notebook/GCD Consultant/zarkowe0/granulosa/res_7d_full_sox9.rnk -scoring_scheme classic -rpt_label sox9 -create_svgs true -include_only_symbols true -make_sets true -plot_top_x 20 -rnd_seed timestamp -set_max 500 -set_min 15 -zip_report false -out /Users/gearh006/Documents/Notebook/GCD Consultant/zarkowe0/granulosa

```

# Figure 6D - Plppr4
```{r plppr4_7d,eval=T}
#as.data.frame(colData(dds_7dko<-dds[,!str_detect(colData(dds)$time,"ovary|testis") & str_detect(colData(dds)$time,"7d|fresh" )  ]))
as.data.frame(colData(dds_7dko<-dds[,str_detect(colData(dds)$time,"7d|fresh" )| (str_detect(colData(dds)$time,"ovary") & str_detect(colData(dds)$gene,"S22_dmrt1|S00_dmrt1|S01_dmrt1") & !str_detect(colData(dds)$batch,"P006") )]))

dds_7dko<-estimateSizeFactors(dds_7dko)
gg_plotCounts7d("Dmrt1",d=dds_7dko,fpkm=TRUE)
gg_plotCounts7d("Sox9",d=dds_7dko,fpkm=TRUE)
gg_plotCounts7d("Sox8",d=dds_7dko,fpkm=TRUE)
gg_plotCounts7d("Plppr4",d=dds_7dko,fpkm=TRUE)

svglite::svglite(paste0("vitro_Plppr4_expression_wKO_",ts,".svg"),width=4.5,height=7)
gg_plotCounts7d("Plppr4",d=dds_7dko,fpkm=TRUE,bar=T)
dev.off()

svglite::svglite(paste0("vitro_Vdr_expression_wKO_",ts,".svg"),width=4.5,height=7)
gg_plotCounts7d("Vdr",d=dds_7dko,fpkm=TRUE,bar=T)
dev.off()
svglite::svglite(paste0("vitro_Syt9_expression_wKO_",ts,".svg"),width=4.5,height=7)
gg_plotCounts7d("Syt9",d=dds_7dko,fpkm=TRUE,bar=T)
dev.off()
svglite::svglite(paste0("vitro_Calr4_expression_wKO_",ts,".svg"),width=4.5,height=7)
gg_plotCounts7d("Calr4",d=dds_7dko,fpkm=TRUE,bar=T)
dev.off()



as.data.frame(colData(dds_7dko_gw<-dds[,str_detect(colData(dds)$batch,"GW" )]))
dds_7dko_gw$gene<-droplevels(dds_7dko_gw$gene)
design(dds_7dko_gw)<-~gene
dds_7dko_gw<-DESeq(dds_7dko_gw)
summary(res_dds_7dko_gw<-results(dds_7dko_gw))
res_dds_7dko_gw<-left_join(rownames_to_column(as.data.frame(res_dds_7dko_gw),var="gene_id"),symbols)
View(res_dds_7dko_gw)


```
